Projekt_MII_Daniel/Projekt_MII_Daniel/requirements.txt
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
reportlab>=4.0
